requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.1
async-timeout==5.0.1
attrs==19.3.0
Automat==0.8.0
bert-score==0.3.13
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==24.4.1
colorama==0.4.3
colorlog==6.10.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cycler==0.12.1
datasets==2.14.7
dbus-python==1.2.16
dill==0.3.7
distro==1.4.0
distro-info==0.23+ubuntu1.1
entrypoints==0.3
fastdtw==0.3.4
filelock==3.16.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2023.10.0
greenlet==3.1.1
hf-xet==1.2.0
holidays==0.58
httplib2==0.14.0
huggingface-hub==0.36.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
importlib_resources==6.4.5
incremental==16.10.1
Jinja2==2.10.1
joblib==1.4.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
langdetect==1.0.9
Mako==1.3.10
MarkupSafe==1.1.0
matplotlib==3.7.5
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
netifaces==0.10.4
networkx==3.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
optuna==4.5.0
packaging==25.0
pandas==2.0.3
pexpect==4.6.0
Pillow==10.4.0
prettytable==0.7.2
propcache==0.2.0
pyarrow==17.0.0
pyarrow-hotfix==0.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
pytz==2025.2
PyYAML==5.3.1
regex==2024.11.6
requests==2.22.0
requests-unixsocket==0.2.0
rouge==1.0.1
rouge-score==0.1.2
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.7.2
SQLAlchemy==2.0.45
ssh-import-id==5.10
sympy==1.13.3
systemd-python==234
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.67.1
transformers==4.35.2
triton==2.1.0
Twisted==18.9.0
typing_extensions==4.13.2
tzdata==2025.3
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
xxhash==3.6.0
yarl==1.15.2
zipp==3.20.2
zope.interface==4.7.1